Move addBackground() from UIView extension to RootViewController but code currently crashes
